refactor is serving into component
