+++ front_end/src/app/admin/events/family-finder/household/household.component.html
@@ -15,6 +15,18 @@
             </div>
           </a>
         </div>
+        <div class="card xl text-xs-center pointer">
+          <label attr.for='isServing' class="text-xs-center pointer px-1 serving-label">
+            <div class="serving-question-cont mt-2">
+              <span class="tgl-inline-label tgl-serving-label">Are you serving today?</span>
+              <input class='tgl tgl-light' id='isServing' type='checkbox' [(ngModel)]="isServing" (click)="toggleClick(showServiceSelectModal)">
+              <label class='tgl-btn' for='isServing'>
+                <span class="tgl-yes">Yes</span>
+                <span class="tgl-no">No</span>
+              </label>
+            </div>
+          </label>
+        </div>
       </div>
       <div *ngIf="eventParticipants && eventParticipants.hasParticipants()">
         <div class="eqWrap mt-2">
