+++ front_end/src/app/child-signin/child-signin.module.ts
@@ -20,6 +20,7 @@ import { MomentModule } from 'angular2-moment';
     AvailableChildrenModule,
     childSigninRouting
   ],
+  exports: []
 })
 
 export class ChildSigninModule { }
